refactor: move query functions inside queries.ts
